feat: take object cmd
