+++ server/Makefile
@@ -23,7 +23,11 @@ SRC			=	src/list/list.c \
 				src/ai/cmd/look/look_east.c \
 				src/ai/cmd/look/look_south.c \
 				src/ai/cmd/look/look_west.c \
-				src/ai/cmd/little.c \
+				src/ai/cmd/other/other_handler.c \
+				src/ai/cmd/other/take_cmd.c \
+				src/ai/cmd/other/fork_cmd.c \
+				src/ai/cmd/other/dead_cmd.c \
+				src/ai/cmd/other/connect_nbr_cmd.c \
 				src/ai/cmd/eject/eject_handler.c \
 				src/ai/cmd/eject/eject_player.c \
 				src/ai/cmd/eject/eject_egg.c \
