+++ server/src/ai/cmd/handler.c
@@ -18,7 +18,7 @@ void command_ai_handler(app_t *app, ia_t *ai, char *line)
         return;
     if (object_info_command(app, ai, line))
         return;
-    if (little_command(app, ai, line))
+    if (other_command(app, ai, line))
         return;
     add_message(ai->list_messages, bad_command);
 }
